+++ src/Kanvas/AccessControlList/Actions/UpdateRoleAction.php
@@ -1,16 +1,16 @@
 <?php
 declare(strict_types=1);
+
 namespace Kanvas\AccessControlList\Actions;
 
+use Illuminate\Auth\Access\AuthorizationException;
 use Kanvas\AccessControlList\Models\Role;
 use Kanvas\AccessControlList\Repositories\RolesRepository;
-use Exception;
-use Illuminate\Auth\Access\AuthorizationException;
 
 class UpdateRoleAction
 {
     /**
-     * __construct
+     * __construct.
      *
      * @return void
      */
@@ -22,15 +22,15 @@ class UpdateRoleAction
     }
 
     /**
-     * execute
+     * execute.
      *
      * @return Role
      */
-    public function execute(): Role
+    public function execute() : Role
     {
         $role = Role::find($this->id);
         if ($role->scope !== RolesRepository::getScope()) {
-            throw new AuthorizationException('You dont have permission to update this role');
+            throw new AuthorizationException('You don\'t have permission to update this role');
         }
         $role->name = $this->name;
         $role->title = $this->title;
