+++ src/Kanvas/SystemModules/Models/SystemModules.php
@@ -9,7 +9,7 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Kanvas\Apps\Models\Apps;
 use Kanvas\Models\BaseModel;
-use Kanvas\Utils\Str;
+use Baka\Support\Str;
 
 /**
  * Apps Model.
